backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nkl708 <nklnsk708@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

django = "*"
django-stubs = "*"
djangorestframework = "*"
djangorestframework-stubs = "*"
djangorestframework-simplejwt = "*"

python-dotenv = "*"
psycopg2-binary = "*"
sentry-sdk = "*"
gunicorn = "*"

[tool.poetry.dev-dependencies]
pylint = "*"
mypy = "*"
bandit = "*"

flake8 = "*"
black = "*"
isort = "*"

pytest = "*"
pytest-django = "*"
pytest-cov = "*"
coverage = "*"
factory-boy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--tb=short --cov=api"

[tool.black]
line-length = 79
string-normalization = true

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "C0114", # Missing module docstring
    "C0115", # Missing class docstring
    "C0116", # Missing function or method docstring
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]

[[tool.mypy.overrides]]
module = "factory.*"
ignore_missing_imports = true
